Ejercicio 10 unidad 3

+++ Programacion/Unidad_3/Ejercicio 10/src/App.java
@@ -3,10 +3,10 @@ public class App {
     public static void main(String[] args) throws Exception {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Introduzca la nota del primer examen:");
-        float nota1 = scanner.nextFloat();
+        double nota1 = scanner.nextFloat();
         System.out.println("Introduzca la nota deseada");
-        float notaFinal = scanner.nextFloat();
-        float nota2 = (notaFinal - (nota1*0.4))/0.6;
+        double notaFinal = scanner.nextFloat();
+        double nota2 = notaFinal - ((nota1*0.4))/0.6;
         System.out.println("La nota necesaria en el segundo para alcanzar un " + notaFinal + " este trimestre teniendo un " + nota1 + " en el primer examen seria: " nota2);
     }
 }
